Add files via upload
I think all the plots are good, I may have forgotten a ylabel  or two but I'm pretty sure they are good to go for the write-up. I'll be including a brief section tomorrow about change over time numerically which I haven't gotten to yet.
Things I found that were notable: summer and spring were peaks for the whole US, while for the W there was a noticeable summer peak, NE and Midwest had peaks in the spring, and the south had a small winter peak. The South and the West combined accounted for 80% of fires even though they are only 54% of states.
I'm gonna commit a new, updated file with the rest of my stat analysis tomorrow before our meeting.